BUGFIX: Corrected mishandling of service querystring parameters.
If the service URL included query string parameters (e.g.
https://service.example.net/?foo=baz), they were improperly encoded during
redirection back to the service such that the values would appear as serialized
Python lists.

This leads to the computed service sent with the ST during validation to *not*
match the service URL stored in CAS with the ST.  An infinite ST request and
validation failure cycle would ensue until the user browser issued a
"too many redirects" error.