Added function to check if URLs are equal beyond a simple string match.
Specifically, URL query string parameters can be rearranged and the resource
identified by the URL *should* be identical.  In practice, this *might* not
be the case, but the CAS service should *probably* be tolerant of this.

I can construct pathalogical examples where this liberal stance could be
problematic, but they all seem to rely on the service provider abusing
the CAS client to its own detriment.

Another subtle match that is sometimes overlooked is when a URL
contains an implicit vs. an explicit port.  E.g. consider the URLS:

    https://www.example.net/
    https://www.example.net:443/

These URLs identify the same resource, but they are different strings.